// src/wasm/trust.d.ts
/* tslint:disable */
/* eslint-disable */

export class Move {
    free(): void;
    [Symbol.dispose](): void;
    constructor(row_offset: number, col_offset: number);
    col_offset: number;
    row_offset: number;
}

export function run_engine(data: Uint8Array, num_rows: number, num_cols: number, player_row: number, player_col: number, opponent_row: number, opponent_col: number): Move;

export function start(): void;

export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembly.Module;

export interface InitOutput {
    readonly memory: WebAssembly.Memory;
    readonly __wbg_get_move_col_offset: (a: number) => number;
    readonly __wbg_get_move_row_offset: (a: number) => number;
    readonly __wbg_move_free: (a: number, b: number) => void;
    readonly __wbg_set_move_col_offset: (a: number, b: number) => void;
    readonly __wbg_set_move_row_offset: (a: number, b: number) => void;
    readonly move_new: (a: number, b: number) => number;
    readonly run_engine: (a: number, b: number, c: number, d: number, e: number, f: number, g: number, h: number) => number;
    readonly start: () => void;
    readonly __wbindgen_free: (a: number, b: number, c: number) => void;
    readonly __wbindgen_malloc: (a: number, b: number) => number;
    readonly __wbindgen_realloc: (a: number, b: number, c: number, d: number) => number;
    readonly __wbindgen_externrefs: WebAssembly.Table;
    readonly __wbindgen_start: () => void;
}

export type SyncInitInput = BufferSource | WebAssembly.Module;

/**
 * Instantiates the given `module`, which can either be bytes or
 * a precompiled `WebAssembly.Module`.
 *
 * @param {{ module: SyncInitInput }} module - Passing `SyncInitInput` directly is deprecated.
 *
 * @returns {InitOutput}
 */
export function initSync(module: { module: SyncInitInput } | SyncInitInput): InitOutput;

/**
 * If `module_or_path` is {RequestInfo} or {URL}, makes a request and
 * for everything else, calls `WebAssembly.instantiate` directly.
 *
 * @param {{ module_or_path: InitInput | Promise<InitInput> }} module_or_path - Passing `InitInput` directly is deprecated.
 *
 * @returns {Promise<InitOutput>}
 */
export default function __wbg_init (module_or_path?: { module_or_path: InitInput | Promise<InitInput> } | InitInput | Promise<InitInput>): Promise<InitOutput>;